Split out a separate class.
