+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/executionengine/ExecutePayloadResult.java
@@ -19,8 +19,6 @@ import java.util.Optional;
 import org.apache.tuweni.bytes.Bytes32;
 
 public class ExecutePayloadResult {
-  public static final ExecutePayloadResult VALID =
-      new ExecutePayloadResult(ExecutionPayloadStatus.VALID, Optional.empty(), Optional.empty());
   private final ExecutionPayloadStatus status;
   private final Optional<Bytes32> latestValidHash;
   private final Optional<String> message;
